# services/api/openapi.json
{
  "components": {
    "schemas": {
      "ApiEnvelope": {
        "additionalProperties": false,
        "description": "Consistent success wrapper. Domain meaning stays inside result.",
        "properties": {
          "api_version": {
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "execution_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Status"
          },
          "input_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Hash"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "product_version": {
            "default": "0.24.3-19",
            "title": "Product Version",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "scientific_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scientific Status"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "result"
        ],
        "title": "ApiEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StructureComplexRequest": {
        "additionalProperties": false,
        "properties": {
          "structure_id": {
            "pattern": "^(4UN3|4OO8)$",
            "title": "Structure Id",
            "type": "string"
          }
        },
        "required": [
          "structure_id"
        ],
        "title": "StructureComplexRequest",
        "type": "object"
      },
      "StructureCoordinateFileRequest": {
        "additionalProperties": false,
        "properties": {
          "structure_id": {
            "pattern": "^(1CRN|1BNA|1RNA)$",
            "title": "Structure Id",
            "type": "string"
          }
        },
        "required": [
          "structure_id"
        ],
        "title": "StructureCoordinateFileRequest",
        "type": "object"
      },
      "StructureFixtureRequest": {
        "additionalProperties": false,
        "properties": {
          "structure_id": {
            "pattern": "^(1CRN|1BNA|1RNA)$",
            "title": "Structure Id",
            "type": "string"
          }
        },
        "required": [
          "structure_id"
        ],
        "title": "StructureFixtureRequest",
        "type": "object"
      },
      "StructureMappedRequest": {
        "additionalProperties": false,
        "properties": {
          "molecule": {
            "default": "PROTEIN",
            "pattern": "^(PROTEIN|DNA|RNA)$",
            "title": "Molecule",
            "type": "string"
          },
          "query_sequence": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Query Sequence",
            "type": "string"
          },
          "structure_id": {
            "pattern": "^(1CRN|1BNA|1RNA)$",
            "title": "Structure Id",
            "type": "string"
          }
        },
        "required": [
          "structure_id",
          "query_sequence"
        ],
        "title": "StructureMappedRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Product version 0.24.3-19. API contract v1. This service validates transport input, calls helixscope_core, and serializes results. It does not independently calculate biology. Job backend: LOCAL MIGRATION JOB BACKEND, NON-DURABLE, NOT FINAL PRODUCTION QUEUE.",
    "summary": "Typed scientific HTTP API for HelixScope core.",
    "title": "HelixScope API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/structures/coordinate-file": {
      "post": {
        "description": "Serves only 1CRN/1BNA/1RNA bundled copies. No user URL.",
        "operationId": "structureCoordinateFile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructureCoordinateFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Allowlisted bundled mmCIF text for the molecular viewer",
        "tags": [
          "structures"
        ]
      }
    },
    "/api/v1/structures/experimental-complex": {
      "post": {
        "description": "4UN3/4OO8 experimental coordinates. Guide mapping stays UNCERTAIN.",
        "operationId": "structureLoadExperimentalComplex",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructureComplexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch a catalogued Cas-guide-DNA complex from RCSB",
        "tags": [
          "structures"
        ]
      }
    },
    "/api/v1/structures/fixture": {
      "post": {
        "description": "Coordinates from the deposited copy. No camera fields.",
        "operationId": "structureLoadFixture",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructureFixtureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Load bundled 1CRN/1BNA/1RNA mmCIF (no network)",
        "tags": [
          "structures"
        ]
      }
    },
    "/api/v1/structures/fixture/mapped": {
      "post": {
        "description": "Preserves mapping_status including UNCERTAIN. No promotion.",
        "operationId": "structureLoadFixtureMapped",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructureMappedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Map a query onto a bundled deposited structure",
        "tags": [
          "structures"
        ]
      }
    },
...
  }
}